setting up fragment to pass back the api object
